requirements.txt
requests==2.17.3
lxml==3.8.0
readability-lxml==0.6.2